sub1 finish

+++ src/Applications/use_case/_test/DeleteCommentUseCase.test.js
@@ -10,31 +10,38 @@ describe('DeleteCommentUseCase', () => {
       commentId: 'comment-xxx',
     };
     const threadUseCasePayload = {
-      threadId: 'thread-xxx'
-    }
-    const fakeOwner = 'owner-xxx'
- 
+      threadId: 'thread-xxx',
+    };
+    const fakeOwner = 'owner-xxx';
+
     /** creating dependency of use case */
     const mockCommentRepository = new CommentRepository();
     const mockThreadRepository = new ThreadRepository();
- 
+
     /** mocking needed function */
+    mockCommentRepository.verifyAvailibilityComment = jest.fn()
+      .mockImplementation(() => Promise.resolve());
+    mockThreadRepository.verifyAvailibilityThread = jest.fn()
+      .mockImplementation(() => Promise.resolve());
+    mockCommentRepository.verifyCommentOwner = jest.fn()
+      .mockImplementation(() => Promise.resolve());
     mockCommentRepository.deleteComment = jest.fn()
-      .mockImplementation(() => Promise.resolve(Promise.resolve()));
-    mockThreadRepository.getThreadById = jest.fn()
-      .mockImplementation(() => Promise.resolve(Promise.resolve()));
- 
+      .mockImplementation(() => Promise.resolve(true));
+
     /** creating use case instance */
     const deleteCommentUseCase = new DeleteCommentUseCase({
       commentRepository: mockCommentRepository,
-      threadRepository: mockThreadRepository
+      threadRepository: mockThreadRepository,
     });
- 
+
     // Action
-    await deleteCommentUseCase.execute(fakeOwner, commentUseCasePayload);
- 
+    const is_delete = await deleteCommentUseCase.execute(fakeOwner, commentUseCasePayload);
+
     // Assert
-    expect(mockThreadRepository.getThreadById).toHaveBeenCalledWith(threadUseCasePayload.threadId);
-    expect(mockCommentRepository.deleteComment).toBeCalledWith(fakeOwner, commentUseCasePayload.commentId);
+    expect(is_delete).toEqual(true);
+    expect(mockThreadRepository.verifyAvailibilityThread).toBeCalledWith(threadUseCasePayload.threadId);
+    expect(mockCommentRepository.verifyAvailibilityComment).toBeCalledWith(commentUseCasePayload.commentId);
+    expect(mockCommentRepository.verifyCommentOwner).toBeCalledWith(fakeOwner, commentUseCasePayload.commentId);
+    expect(mockCommentRepository.deleteComment).toBeCalledWith(commentUseCasePayload.commentId);
   });
 });
